requirements.txt
Django>=3.1.5
webstack-django-jwt-auth>=1.3.0
httpx>=0.16.1
